refactor: split VmfFile implementation into modules
